Updating documentation of `pyttb.tensor.from_function`. (#425)
* Updating documentation of `pyttb.tensor.from_function`.
Added an example showing how to create a tensor with normal random entries, for which there is no current shortcut like `tenones`.
Used that example to show how it's arguably best to pass in a vector rather than trying to worry about forcing Fortran ordering.
Also updated the example of create an all ones tensor.
The previous documentation relied on some function called `fortran_order_one`, and I'm not sure what that is.
Instead, I changed it to the `numpy.ones` function and to specify the fortran order via its interface.
I think these examples will be more helpful to users.

* Fixing ruff error.

---------

Co-authored-by: Danny Dunlavy <dmdunla@sandia.gov>